Add nullability annotations
This is a new language feature, so there are not established best practices yet. I have not used assume non-null blocks, since explicit is generally better than adding another layer of state needed to understand the code.

It seems unnecessary to check a ‘nonnull’ pointer for nil so these were removed. This might turn out to be a bad idea since it leans on the compiler to catch pointers that might be nil, and it probably will not be able to do this because some other code might not have nullability annotations.

Some code has not been annotated, because it isn’t clear how this should be done. This is mostly because of dependency on Apple frameworks, which are not fully or clearly annotated. (You can get an idea by looking at the Swift ‘headers’).